Some test fixes.
