+++ openmdao/core/system.py
@@ -1,6 +1,7 @@
 """Define the base System class."""
 import sys
 import os
+import hashlib
 import time
 
 from contextlib import contextmanager
@@ -5058,3 +5059,46 @@ class System(object):
         relevant['nonlinear'] = relevant['linear']
 
         return relevant
+
+    def _generate_md5_hash(self):
+        """
+        Generate an md5 hash for the data structure of this model.
+
+        The hash is generated from an encoded string containing the physical model hiearchy,
+        including all component and variable names, and all connection information.
+
+        The hash is used by the n2 viewer to determine if a saved view can be reused. It is not
+        intended to accurate track whether a model has been changed, so no options/settings are
+        tracked.
+
+        Returns
+        -------
+        str
+            The md5 hash string for the model.
+        """
+        data = []
+
+        # Model Hierarchy.
+        for sys_name in self.system_iter(include_self=True, recurse=True):
+
+            # System name and depth.
+            pathname = sys_name.pathname
+            if pathname:
+                name_parts = pathname.split('.')
+                depth = len(name_parts)
+
+                data.append((name_parts[-1], depth))
+
+            else:
+                data.append(('model', 0))
+
+            # Local (relative) names for Component inputs and outputs.
+            try:
+                data.append(sys_name._var_rel_names)
+            except AttributeError:
+                continue
+
+        # All Connections.
+        data.append(self._conn_global_abs_in2out)
+
+        return hashlib.md5(str(data).encode()).hexdigest()
