+++ openmdao/recorders/tests/test_sqlite_reader.py
@@ -2104,7 +2104,7 @@ class TestSqliteCaseReader(unittest.TestCase):
 
         self.assertEqual(set(cr.problem_metadata.keys()), {
             'tree', 'sys_pathnames_list', 'connections_list', 'variables', 'abs2prom',
-            'driver', 'design_vars', 'responses', 'declare_partials_list'
+            'driver', 'design_vars', 'responses', 'declare_partials_list', 'md5_hash'
         })
 
         self.assertEqual(len(cr._driver_cases._cases), 0)
@@ -2132,7 +2132,7 @@ class TestSqliteCaseReader(unittest.TestCase):
 
         self.assertEqual(set(cr.problem_metadata.keys()), {
             'tree', 'sys_pathnames_list', 'connections_list', 'variables', 'abs2prom',
-            'driver', 'design_vars', 'responses', 'declare_partials_list'
+            'driver', 'design_vars', 'responses', 'declare_partials_list', 'md5_hash'
         })
 
         self.assertEqual(len(cr._driver_cases._cases), num_driver_cases)
